third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/inline/iterable.tentative-expected.txt

This is a testharness.js-based test.
PASS Iterating over an empty StylePropertyMap gives a zero-length array
PASS StylePropertyMap iterates properties in inline style order
PASS StylePropertyMap iterator returns CSS properties with the correct CSSStyleValue
PASS StylePropertyMap iterator returns list-valued properties with the correct CSSStyleValue
FAIL StylePropertyMap iterator returns custom properties with the correct CSSStyleValue assert_equals: expected "A" but got " A"
Harness: the test ran to completion.
